fixed breaking WIP, made encoded_dtype configurable - helps with repeat

+++ shape_tfds/core/features/voxel.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
+import tensorflow as tf
 import shape_tfds as sds
 
 
@@ -18,6 +19,7 @@ def BinaryVoxel(shape):
         raise ValueError('shape must be a tuple, got %s' % str(shape))
     return sds.core.features.PaddedTensor(
         sds.core.features.ShapedTensor(
-            sds.core.features.BinaryRunLengthEncodedFeature(),
+            sds.core.features.BinaryRunLengthEncodedFeature(
+                serialized_dtype=tf.uint8),
             (None,)*len(shape)),
         padded_shape=(shape))
